Move state managing logic from proxies to qt
