+++ yt/yt/server/lib/discovery_server/unittests/mock/connection.h
@@ -198,6 +198,11 @@ public:
         YT_UNIMPLEMENTED();
     }
 
+    NRpc::IChannelPtr GetQueryTrackerChannelOrThrow(const TString& /*stage*/) override
+    {
+        YT_UNIMPLEMENTED();
+    }
+
     const NTabletClient::ITableMountCachePtr& GetTableMountCache() override
     {
         YT_UNIMPLEMENTED();
