+++ yt/yt/server/query_tracker/private.h
@@ -39,6 +39,7 @@ YT_DEFINE_ERROR_ENUM(
 ////////////////////////////////////////////////////////////////////////////////
 
 DECLARE_REFCOUNTED_CLASS(TQueryTracker)
+DECLARE_REFCOUNTED_CLASS(TQueryTrackerProxy)
 DECLARE_REFCOUNTED_CLASS(TQueryTrackerDynamicConfig)
 DECLARE_REFCOUNTED_CLASS(TQueryTrackerServerConfig)
 DECLARE_REFCOUNTED_CLASS(TQueryTrackerServerDynamicConfig)
