+++ yt/yt/server/query_tracker/proxy_service.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "private.h"
+
+#include <yt/yt/core/rpc/public.h>
+
+namespace NYT::NQueryTracker {
+
+////////////////////////////////////////////////////////////////////////////////
+
+NRpc::IServicePtr CreateProxyService(IInvokerPtr proxyInvoker, TQueryTrackerProxyPtr proxy);
+
+////////////////////////////////////////////////////////////////////////////////
+
+} // namespace NYT::NQueryTracker
