+++ yt/yt/tests/conftest_lib/conftest_queries.py
@@ -45,8 +45,22 @@ class QueryTracker(ExternalComponent):
                          verbose=False, verbose_error=False),
              ignore_exceptions=True)
 
+    def on_start(self):
+        query_tracker_config = {
+            "stages": {
+                "production": {
+                    "channel": {
+                        "addresses": self.addresses,
+                    }
+                },
+            },
+        }
+        set("//sys/clusters/primary/query_tracker", query_tracker_config)
+        wait(query_tracker_has_loaded)
+
     def on_finish(self):
         remove("//sys/query_tracker/instances", recursive=True, force=True)
+        remove("//sys/clusters/primary/query_tracker")
 
 
 def query_tracker_has_loaded():
@@ -61,21 +75,13 @@ def query_tracker_environment():
     create_user("query_tracker")
     add_member("query_tracker", "superusers")
     sync_create_cells(1)
-    query_tracker_config = {
-        "stages": {
-            "production": {},
-        },
-    }
-    set("//sys/clusters/primary/query_tracker", query_tracker_config)
+
     create("document", "//sys/query_tracker/config", recursive=True, force=True, attributes={"value": {}})
     create_access_control_object_namespace("queries")
     create_access_control_object("nobody", "queries")
-
-    wait(query_tracker_has_loaded)
     yield
     remove("//sys/access_control_object_namespaces/queries/nobody")
     remove("//sys/access_control_object_namespaces/queries")
-    remove("//sys/clusters/primary/query_tracker")
     sync_remove_tablet_cells(ls("//sys/tablet_cells"))
     remove_user("query_tracker")
     remove("//sys/query_tracker", recursive=True, force=True)
