+++ yt/yt/ytlib/api/native/connection.h
@@ -83,6 +83,7 @@ struct IConnection
     virtual const NChunkClient::IChunkReplicaCachePtr& GetChunkReplicaCache() = 0;
 
     virtual std::pair<IClientPtr, NYPath::TYPath> GetQueryTrackerStage(const TString& stage) = 0;
+    virtual NRpc::IChannelPtr GetQueryTrackerChannelOrThrow(const TString& stage) = 0;
 
     virtual const NHiveClient::TCellTrackerPtr& GetDownedCellTracker() = 0;
