+++ yt/yt/ytlib/query_tracker_client/config.h
@@ -2,9 +2,26 @@
 
 #include "public.h"
 
+#include <yt/yt/core/rpc/config.h>
+
 #include <yt/yt/core/ytree/yson_struct.h>
 
 namespace NYT::NQueryTrackerClient {
+
+///////////////////////////////////////////////////////////////////////////////
+
+class TQueryTrackerChannelConfig
+    : public NRpc::TBalancingChannelConfig
+{
+public:
+    TDuration Timeout;
+
+    REGISTER_YSON_STRUCT(TQueryTrackerChannelConfig);
+
+    static void Register(TRegistrar registrar);
+};
+
+DEFINE_REFCOUNTED_TYPE(TQueryTrackerChannelConfig)
 
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -12,6 +29,7 @@ class TQueryTrackerStageConfig
     : public NYTree::TYsonStruct
 {
 public:
+    TQueryTrackerChannelConfigPtr Channel;
     NYPath::TYPath Root;
     TString User;
