+++ yt/yt/ytlib/query_tracker_client/helpers.h
@@ -2,6 +2,10 @@
 
 #include "public.h"
 
+#include <yt/yt/client/api/query_tracker_client.h>
+
+#include <yt/yt/ytlib/query_tracker_client/proto/query_tracker_service.pb.h>
+
 namespace NYT::NQueryTrackerClient {
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -16,5 +20,8 @@ bool IsPreFinishedState(EQueryState state);
 bool IsFinishedState(EQueryState state);
 
 ////////////////////////////////////////////////////////////////////////////////
+
+void ToProto(NProto::TQuery* protoQuery, const NApi::TQuery& query);
+void FromProto(NApi::TQuery* query, const NProto::TQuery& protoQuery);
 
 } // namespace NYT::NQueryTrackerClient
